Stokhos:  Fix serialization for UQ::PCE and add unit tests.
The serialization object for UQ::PCE didn't compile or function
correctly.  I fixed this and added unit tests for it.  Part of the
problem is I need to remember not to initialize PCE objects like this:

PCE x;
x = PCE(cijk);
for (int i=0; i<cijk.dimension(); ++i)
  x.fastAccessCoeff(i) = ...

The reaons is the assignment operator checks is the RHS is constant (all
terms but the first zero), and x in this case will be constant too,
resulting in invalid writes in the subsequent loop.  Ultimately I would
like to remove that logic, but is currently necessary to get the
mean-based preconditioner setup to be efficient in MueLu.  Instead we
need to do:

PCE x;
x.reset(cijk);
for (int i=0; i<cijk.dimension(); ++i)
  x.fastAccessCoeff(i) = ...