Moertel: Fix a bug related to Newton solution of projection equation.
The equation expressing the projection of a master node onto a nonmaster segment
based on nonmaster node normals is a nonlinear equation in one
variable. Newton's method is used to solve this equation.

Previously, the function implementing the Newton's method for this equation did
not return a failure status. In this context, Newton's method fails almostly
certainly only if there is no solution to the projection equation. Hence failure
is an important piece of information.

In addition, the success status for the other associated equation was ignored in
the driving routine.

The result of not using these two success/fail data was that spurious supposed
solutions were being used in subsequent calculations when, by a confluence of
factors, the spurious solutions happened to fall in the range of allowed values
filtered at a higher level. This led to spikes in the gap functions at almost
arbitrarily small scales in some geometries. In turn, this meant that at the
scale of a step in an implicit method near convergence, the gap functions could
be highly discontinuous, preventing convergence.

The bug was not always active. In my simulations, I solve a sequence of contact
problems with continuation on the BCs. I found that my solver was having
convergence issues in perhaps 5-10% of the posed problems. These failures were
due to this bug.