Tpetra: Bug 6051 test actually should have passed; now it does.
Eric Phipps dug into the Teuchos unit test framework issue, and
noticed that the Bug 6051 test should have been failing, even though
it was not failing.  This is because the Teuchos unit test framework
was only looking at test results from Process 0, even though all
processes were checking things.  When Eric added options to make the
unit test framework check all processes, the test failed.  I verified
that it failed for me also, when I added those options.

I took a closer look at the test and noticed that its idea of
"passing" depended on the order of global indices in the result of
Tpetra::createOneToOne.  That function does not promise a particular
order, esp. when using a custom TieBreak function.  This commit
changes the test so that it checks global indices, independently of
order.  It also adds the options to force the Teuchos unit test
framework to check results on all processes.  Now the test passes, and
justifiably so.