Tpetra: Fixed Bug 5822 test (Map_Bug5822_2).
I realize now that the test might have been failing unjustifiedly,
when neither of GlobalOrdinal = long or GlobalOrdinal = long long were
enabled.  This commit fixes that.  The test is still only meaningful
when GlobalOrdinal is a 64-bit type, but now the test passes with an
early exit and an informative message if no 64-bit GlobalOrdinal types
are enabled.  (Now it tests the right macros -- not the Teuchos ones
but the Tpetra ones.)