STK create_edges bugfix
Fix Ticket #13442

Problem was that create_edges does not hook edges to faces not in the selector locally, but the modification end
cleanup step does hook edges to ghosted faces even if they are not in the selector passed to create_edges()

Fix is to hook edges to any face, even if that face is not in the selector passed to create_edges(). This
emulates the create_faces() and I/O behavior of create_faces in Framework.

Paired with Dave Glaze and sometimes Todd Coffey and Tolu Okunsanya.

(From Sierra)