Tpetra: Added new function Details::gathervPrint.
Tpetra::Details::gathervPrint takes a string from each process in a
given communicator.  It then prints the strings in rank order, to an
output stream on Process 0 of that communicator.

I wrote this to help with debugging Tpetra, with respect to the
"feature" of the Teuchos test framework, that only Process 0 prints
anything (unless you tell it otherwise).  However, it may be generally
useful for other situations in which only Process 0 is allowed to
print.  For example, Matrix Market output for Tpetra::MultiVector uses
something like this, so we could consolidate code.