Stokhos:  Replace AnalyzeSacadoShape with AnalyzeStokhosShape in UQ types
The Sacado and Stokhos View specializations use AnalyzeSacadoShape to
determine the scalar type layout (similar to Kokkos' AnalyzeShape with
some modifications).  This was causing a problem when nesting MP::Vector
inside of Fad when trying to use views of rank 7.  Since the UQ types
are considered to be atomic, they don't need to expose their
inner-dimensions to the Fad specialization.  So I replaced the usage in
Stokhos with AnalyzeStokhosShape (which already existed but wasn't
used).  I also added a unit test that creates a rank-7 view to verify
this no works.