MueLu: fixing ETI logic for corner cases
If Tpetra is disabled, disable MueLu ETI support, since it depends on
Tpetra for ETI definitions.  Tpetra might be disabled because CXX11
support is off, or it might be explicitly disabled.

This fix is a little more involved because one might still have asked for
Trilinos_ENABLE_EXPLICIT_INSTANTATION=ON, in which case MueLu has to
turn off some internal ETI cmake variables.  In this case, it's
valid to build the MueLu tests and examples.  Previously, this was
failing at configure time.