Domi: Fixed some const-correctness issues
When I used const Domi::MDArrayView< int > as an argument (for
processor dimensions or map dimensions), I was saying tht the data
(int) was non-const and therefore mutable.  My SWIG typemaps look at
this and have to assume that the values can be changed in-place, and
so disallows Python tuples or lists as Python substitutes, requiring a
NumPy array.

By making these arguments const Domi::MDArrayView< const int >, my
SWIG typemaps know they are purely input data that won't be changed
in-place, and allows Python tuples and lists as substitutes.  This
makes for cleaner, and more correct, Python code (as well as C++
code).